requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
langchain-openai
psycopg2
pinecone-client